+++ src/main/java/br/edu/ifpb/pautax/application/useCases/administrador/alunos/deletar/IDeletarAlunoUseCase.java
@@ -1,5 +1,5 @@
 package br.edu.ifpb.pautax.application.useCases.administrador.alunos.deletar;
 
 public interface IDeletarAlunoUseCase {
-    String execute(Long id);
+    String execute(int id);
 }
