+++ src/main/java/br/edu/ifpb/pautax/application/useCases/administrador/alunos/editar/IEditarAlunoUseCase.java
@@ -3,5 +3,5 @@ package br.edu.ifpb.pautax.application.useCases.administrador.alunos.editar;
 import org.springframework.web.servlet.ModelAndView;
 
 public interface IEditarAlunoUseCase {
-    ModelAndView execute(Long id);
+    ModelAndView execute(int id);
 }
